DCS-19: Using new filenames instead of old.
