+++ src/DiplomContentSystem.Services/IQueryBuilder.cs
