+++ src/DiplomContentSystem.Services/QueryBuilderBase.cs
@@ -1,15 +1,16 @@
 using System;
+using System.Linq;
 using System.Linq.Expressions;
 using DiplomContentSystem.Core;
 using DiplomContentSystem.Dto;
 
 namespace DiplomContentSystem.Services
 {
-    public abstract class QueryBuilder<T> : IQueryBuilder<T> where T : class, Core.IEntity
+    public abstract class QueryBuilderBase<T> : IQueryBuilder<T> where T : class, Core.IEntity
     {
         protected Dto.Request _requestDto;
         protected Query<T> _dbQuery;
-        public QueryBuilder()
+        public QueryBuilderBase()
         {
             _dbQuery = new Query<T>();
         }
@@ -25,8 +26,23 @@ namespace DiplomContentSystem.Services
         }
         public virtual IQueryBuilder<T> UseSortings(string defaultSorting)
         {
+            _dbQuery.SortExpressions = new System.Collections.Generic.List<SortExpression<T>>();
+            if (_requestDto.Sortings == null || !_requestDto.Sortings.Any())
+            {
+                _dbQuery.SortExpressions.Add(new SortExpression<T>(GetSortExpression(defaultSorting), SortDirection.Ascending));
+            }
+            else
+            {
+                foreach (var sorting in _requestDto.Sortings)
+                {
+                    _dbQuery.SortExpressions.Add(new SortExpression<T>(GetSortExpression(sorting.FieldName), (SortDirection)sorting.Direction));
+                }
+            }
             return this;
         }
+
+        protected abstract Expression<Func<T, object>> GetSortExpression(string sortParamName);
+
         public IQueryBuilder<T> UsePaging()
         {
             _dbQuery.Skip = _requestDto.Skip;
